Embed sample sqp client
