+++ lib/svrsample/common/interfaces.go
@@ -1,0 +1,17 @@
+package common
+
+// QueryResponder represents an interface to a concrete type which responds
+// to query requests.
+type QueryResponder interface {
+	Respond(clientAddress string, buf []byte) ([]byte, error)
+}
+
+// QueryState represents the state of a currently running game.
+type QueryState struct {
+	CurrentPlayers int32
+	MaxPlayers     int32
+	ServerName     string
+	GameType       string
+	Map            string
+	Port           uint16
+}
