--- lib/svrsample/protocol/sqp/register.go
@@ -1,7 +1,0 @@
-package sqp
-
-import "github.com/multiplay/go-svrquery/lib/svrsample/protocol"
-
-func init() {
-	protocol.MustRegister("sqp", NewQueryResponder)
-}
